Added diagnostic startup display
The Ognite display will now show a short test pattern on power-up to
help you debug any connections problems.

The pattern starts by filling in the whole screen with on LEDs
left-to-right then bottom-to-top. It then repeats the same path turning
all LEDs off as it goes. Once that is done, it goes on to the normal
flame display.

Also cleaned up some types, making framecounttype  which currently a
byte but can be easily changed to a word if we can fit more than 256
frames.

Made a macro to access the gamma correction table from Flash. Of course
I knew I should do this when I first wrote it, but didn't, and then it
made a bug when I made some other changes

Got rid of all the Ardunio defines since there is really no going back
at this point.

Sped up the getnextbit slightly with static registers, but this should
really be ASM.

Also sped up some bit handling in playvideo().